changed charge from int to float to handle quarks

+++ interface/TGenParticle.h
@@ -18,7 +18,7 @@ public:
   int status;
   int id;
   int motherId;
-  int charge;
+  float charge;
   
   TLorentzVector lv;
   void setLV(){
